wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3965 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/fitalgorithm_fit_gsl/qffitalgorithmgslderivbase.cpp
@@ -72,7 +72,7 @@ QFFitAlgorithm::FitResult QFFitAlgorithmGSLDerivativeBase::intFit(double* params
 
     int iter = 0;
     int maxIter = getParameter("max_iterations").toInt();
-    double gtol = getParameter("gtol").toInt();
+    double gtol = getParameter("gtol").toDouble();
     int status;
 
     ;
